+++ packages/client/src/pages/Forum/components/BlockComments/BlockComment.module.scss
@@ -42,25 +42,4 @@
   flex-flow: column;
 }
 
-.topic {
-  display: flex;
-  flex-flow: row;
-  gap: 10px;
-
-  .text {
-    @include fontTitle(normal, 400, 14px, 20px);
-    color: $text-primary;
-    margin: 5px 10px;
-  }
-
-  .input {
-    flex: 1;
-  }
-
-  .button_comment {
-    @include fontText(normal, 400, 14px, 20px);
-    padding: 0px 20px;
-    height: auto;
-  }
-}
