+++ packages/client/src/pages/Forum/components/BlockComments/BlockComments.tsx
@@ -1,45 +1,19 @@
-import React, { ChangeEvent, useCallback, useState } from 'react';
+import React, { useCallback } from 'react';
 import { BlankWindow } from 'src/components/BlankWindow';
 import styles from './BlockComment.module.scss';
 import { Topic } from '../../part/Topic';
 import { IComment } from '../../part/Comment/types';
 import { Comment } from '../../part/Comment';
-import { Input } from 'src/components/Input';
 import { Button } from 'src/components/Button';
-import { addCommentInTopic } from 'src/store/forum/ForumSlice';
-import { dateFormatting } from 'src/utils/dateFormatting';
-import { useAppDispatch, useAppSelector } from 'src/hooks/redux';
 import { useNavigate } from 'react-router-dom';
 import { ErrorBoundary } from 'src/components/ErrorBoundary';
+import { useAppSelector } from 'src/hooks/redux';
+import { SendComment } from '../../part/SendComment';
 
 export const BlockComments = () => {
   const navigate = useNavigate();
-  const [textComment, setTextComment] = useState('');
-  const [login, topic] = useAppSelector(state => [state.auth.user?.login, state.forum.activeTopic]);
-  const dispatch = useAppDispatch();
 
-  const changeComment = useCallback((event: ChangeEvent) => {
-    const input = event.target as HTMLInputElement;
-    setTextComment(input.value);
-  }, []);
-
-  const addComment = useCallback((): void => {
-    const newFormatDate = dateFormatting(new Date());
-    if (topic) {
-      dispatch(
-        addCommentInTopic({
-          id: topic.id,
-          comment: {
-            text: textComment,
-            author: login || '',
-            date: newFormatDate,
-            likes: 0,
-          },
-        })
-      );
-    }
-    setTextComment('');
-  }, [dispatch, topic, login, textComment]);
+  const [topic] = useAppSelector(state => [state.forum.activeTopic]);
 
   const handlerCreateTopic = useCallback(() => {
     navigate('/forum/createtopic');
@@ -58,18 +32,7 @@ export const BlockComments = () => {
             })}
           </div>
           <BlankWindow className={styles.card}>
-            <div className={styles.topic}>
-              <div className={styles.text}>New comment: </div>
-              <Input
-                name="comment"
-                className={styles.input}
-                onChange={changeComment}
-                value={textComment}
-              />
-              <Button regular className={styles.button_comment} onClick={addComment}>
-                Send
-              </Button>
-            </div>
+            <SendComment />
           </BlankWindow>
         </div>
         }
