+++ packages/client/src/pages/Forum/components/BlockCreateTopic/CreateTopic.module.scss
@@ -24,7 +24,7 @@
   box-sizing: border-box;
 }
 
-.topic {
+.new_topic {
   display: flex;
   flex-flow: row;
   margin-top: 10px;
