+++ packages/client/src/pages/Forum/part/SendComment/SendComment.module.scss
@@ -1,0 +1,23 @@
+@import 'src/index.module.scss';
+
+.new_comment {
+    display: flex;
+    flex-flow: row;
+    gap: 10px;
+  
+    .text {
+      @include fontTitle(normal, 400, 14px, 20px);
+      color: $text-primary;
+      margin: 5px 10px;
+    }
+  
+    .input {
+      flex: 1;
+    }
+  
+    .button_comment {
+      @include fontText(normal, 400, 14px, 20px);
+      padding: 0px 20px;
+      height: auto;
+    }
+  }
