+++ packages/client/src/pages/Forum/part/SendComment/index.ts
@@ -1,0 +1,1 @@
+export { SendComment } from './SendComment';
