+++ packages/client/src/store/forum/initialSlice.ts
@@ -1,6 +1,6 @@
 import { ForumState } from '../types';
 import topics from 'src/static/topics.json';
-import { ITopic } from 'src/pages/Forum/components/Topic/types';
+import { ITopic } from 'src/pages/Forum/part/Topic/types';
 
 export const initialState: ForumState = {
   activeTopic: undefined,
